Fix #14240: Actions show blank link for themes
When clicking to open the list dropdown for a theme on the Themes page,
blank links are shown at the top and it happens only for themes that are
not set as system theme.
To fix this, Resources/views/Theme/list.html.twig was updated to prevent
empty button definitions from being added to the `customButtons` array.
Now, only buttons with proper attributes are merged into the list of
actions shown in the dropdown menu.
An E2E test is included to verify that no links with
`data-toggle="ajax"` are rendered without a `href` attribute.